move extractDate helper into helper/controller
