+++ app/helpers/generateErrorSummary.js
@@ -1,0 +1,6 @@
+export const generateErrorSummary = ({ errorMap }) => (
+  Object.entries(errorMap).map(([questionId, errors]) => ({
+    href: `#${questionId}`,
+    text: errors.errorMessages.join(', '),
+  }))
+);
